[Feat] #56 책톡:채팅 키보드 제어 - ChatInputWrap widget으로 별도 작성 - 외부 클릭시 키보드, 첨부파일 화면 hide - 채팅 input 클릭시 키보드 표시 - 첨부파일 버튼 클릭시 키보드 hide - 첨부파일 화면 표시 중 채팅 input 클릭시 첨부파일 hide, 키보드 show
