+++ lib/modules/chat/view/screens/book_talk_chat_room_screen.dart
@@ -1,5 +1,4 @@
 import 'package:ably_flutter/ably_flutter.dart' as ably;
-import 'package:book/common/components/text_field/search_text_field.dart';
 import 'package:book/common/theme/style/app_paddings.dart';
 import 'package:book/common/theme/style/app_texts.dart';
 import 'package:book/gen/assets.gen.dart';
@@ -9,6 +8,7 @@ import 'package:book/modules/auth/view_model/auth_view_model.dart';
 import 'package:book/modules/chat/model/chat_message_request.dart';
 import 'package:book/modules/chat/model/chat_message_response.dart';
 import 'package:book/modules/chat/state/chat_state.dart';
+import 'package:book/modules/chat/view/widgets/chat_input_wrap.dart';
 import 'package:book/modules/chat/view_model/chat_view_model.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:collection/collection.dart';
@@ -34,6 +34,7 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
   bool _visibleOption = false;
   final ImagePicker _picker = ImagePicker();
   ably.RealtimeChannel? _channel;
+  final FocusNode chatInputFocusNode = FocusNode();
 
   @override
   void initState() {
@@ -81,6 +82,7 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
     setState(() {
       _visibleOption = value;
     });
+    if (value) chatInputFocusNode.unfocus();
   }
 
   _handleTextSend() async {
@@ -124,6 +126,11 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
     }
   }
 
+  _clearChatInput() {
+    chatInputFocusNode.unfocus();
+    _updateVisibleOption(false);
+  }
+
   @override
   void dispose() {
     _channel?.detach();
@@ -141,28 +148,39 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
             orElse: () => ChatViewModel.defaultCategory)
         .name;
 
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(roomName),
-        actions: [
-          IconButton(
-              icon: const Icon(Icons.menu),
-              onPressed: () =>
-                  context.push('/book-talk/chat-room/${widget.roomId}/menu'))
-        ],
-      ),
-      body: SafeArea(
-        child: state.when(
-          data: (data) => Column(
-            children: [
-              Expanded(
-                  child: _buildChatHistory(_scrollController, data, memberId)),
-              _buildChatInputWrap(_textController, _visibleOption,
-                  _updateVisibleOption, _handleTextSend, _clickInputOption)
-            ],
+    return GestureDetector(
+      onTap: () {
+        _clearChatInput();
+      },
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text(roomName),
+          actions: [
+            IconButton(
+                icon: const Icon(Icons.menu),
+                onPressed: () =>
+                    context.push('/book-talk/chat-room/${widget.roomId}/menu'))
+          ],
+        ),
+        body: SafeArea(
+          child: state.when(
+            data: (data) => Column(
+              children: [
+                Expanded(
+                    child: _buildChatHistory(_scrollController, data, memberId)),
+                ChatInputWrap(
+                  textController: _textController,
+                  visibleOption: _visibleOption,
+                  focusNode: chatInputFocusNode,
+                  updateVisibleOption: _updateVisibleOption,
+                  handleTextSend: _handleTextSend,
+                  clickInputOption: _clickInputOption,
+                )
+              ],
+            ),
+            loading: () => const Center(child: CircularProgressIndicator()),
+            error: (err, stack) => Center(child: Text('Error: $err')),
           ),
-          loading: () => const Center(child: CircularProgressIndicator()),
-          error: (err, stack) => Center(child: Text('Error: $err')),
         ),
       ),
     );
@@ -270,12 +288,16 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
                                     child: getChatWidget(message, textStyle),
                                   ),
                                 ),
-                                SizedBox(width: 6,),
+                                SizedBox(
+                                  width: 6,
+                                ),
                                 getDateWidget(message.createdAt),
                               ]
                             : [
                                 getDateWidget(message.createdAt),
-                                SizedBox(width: 6,),
+                                SizedBox(
+                                  width: 6,
+                                ),
                                 ConstrainedBox(
                                   constraints: BoxConstraints(maxWidth: 140),
                                   child: Container(
@@ -313,142 +335,6 @@ class _BookTalkChatRoomScreen extends ConsumerState<BookTalkChatRoomScreen> {
             "채팅을 시작해 보세요",
             style: AppTexts.b8.copyWith(color: ColorName.g3),
           ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildChatInputWrap(
-      TextEditingController textController,
-      bool visibleOption,
-      Function(bool) updateVisibleOption,
-      Function() handleTextSend,
-      Function(ImageSource) clickInputOption) {
-    return Container(
-      height: !visibleOption ? 129 : 129 + 249,
-      decoration: BoxDecoration(
-        color: ColorName.g7,
-        borderRadius: BorderRadius.only(
-          topLeft: Radius.circular(20),
-          topRight: Radius.circular(20),
-        ),
-      ),
-      child: Padding(
-        padding: AppPaddings.CHAT_INPUT_PADDING,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            _buildChatInput(
-              textController,
-              visibleOption,
-              updateVisibleOption,
-              handleTextSend,
-            ),
-            if (visibleOption) _buildChatInputOption(clickInputOption)
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildChatInput(
-      TextEditingController textController,
-      bool visibleOption,
-      Function(bool) updateVisibleOption,
-      Function() handleTextSend) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        GestureDetector(
-          child: !visibleOption
-              ? Assets.icons.icBookpickChatOption.svg()
-              : Icon(Icons.close),
-          onTap: () {
-            updateVisibleOption(!visibleOption);
-          },
-        ),
-        SizedBox(
-          width: 24,
-        ),
-        Expanded(
-          child: SearchTextField(
-            controller: textController,
-            backgroundColor: ColorName.w1,
-            textColor: ColorName.b1,
-            hintText: '메세지를 작성해 보세요',
-            hintStyle: AppTexts.b8.copyWith(color: ColorName.g3),
-            suffixIcon: Padding(
-              padding: const EdgeInsets.symmetric(vertical: 13),
-              child: textController.text.isNotEmpty
-                  ? Assets.icons.icBookpickChatSendColored
-                      .svg(width: 22, height: 22)
-                  : Assets.icons.icBookpickChatSendDisabled
-                      .svg(width: 22, height: 22),
-            ),
-            onTap: () {
-              updateVisibleOption(false);
-            },
-            onTapSuffixIcon: () {
-              handleTextSend();
-            },
-            keyboardType: TextInputType.multiline,
-            maxLines: null,
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildChatInputOption(Function(ImageSource) clickInputOption) {
-    return Expanded(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              _buildOptionButton(
-                  "카메라", Assets.icons.icCamera.svg(width: 27, height: 27), () {
-                clickInputOption(ImageSource.camera);
-              }),
-              _buildOptionButton(
-                  "갤러리",
-                  Assets.icons.icBookpickChatOptionPicture
-                      .svg(width: 24, height: 24), () {
-                clickInputOption(ImageSource.gallery);
-              })
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildOptionButton(
-    String text,
-    Widget icon,
-    Function() onTap,
-  ) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Column(
-        children: [
-          Container(
-              width: 55,
-              height: 55,
-              decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(38.5),
-                  color: ColorName.dim2),
-              child: Center(child: icon)),
-          SizedBox(
-            height: 6,
-          ),
-          Text(
-            text,
-            style: AppTexts.b8.copyWith(color: ColorName.g2),
-          )
         ],
       ),
     );
